refactor(api): delete root UnifiedHandlers; collapse construction onto SharedServices (TD-104 S3-f)
Terminal TD-104 step: stop constructing the legacy root `UnifiedHandlers`
and delete it (src/api_handlers/request_handlers.rs, 4,273 lines, + its
tests). Every network/embedded surface now routes through the runtime
`proximadb_runtime::UnifiedHandlers` port handler; the record write path
lives in `RecordOpsService`.

Construction collapse (shared_services.rs):
- Build the shared `Arc<RecordOpsService>` once (was inside ROOT::new); add
  it + `graph_collection_service` as `SharedServices` fields so every
  former `request_handlers.record_ops()` / `.graph_collection_service` site
  shares one Arc.
- Relocate ROOT's post-construction setters onto their real owners:
  set_dml_service / set_lease_manager / set_precision_resolver ->
  record_ops; drop set_query_adapter (read only inside ROOT) and
  set_hybrid_runtime (resolve_hybrid_static was ROOT-internal, dead).
- Drop the `request_handlers` field.

Relocation (record_ops_service.rs):
- `CollectionIdCache` (+entry, Default) and `enforce_wal_lane_for_record_batch`
  move here from the deleted module (their only surviving consumer).
- Remove the now-vestigial DdlService field/setter/getter (DDL routes via
  the runtime handler's QueryFacadeAdapter, not the record path).

Callers:
- `RestCoreServices::from_shared_services` reads the new SharedServices fields.
- multi_server / database use `services.record_ops`; embedded routes
  `execute_sql_v1` through `api_handlers` (SqlValue params lowered to
  ProximaValue at the boundary, matching REST).

Verified: cargo check/clippy --lib --bins -D warnings, rustup 1.95.0 fmt,
server bin build (default + cluster feature), 8417 unit tests pass, and all
deterministic commit guards (docs-claim, capability-matrix,
workspace-boundaries, tenant-path, panic-policy) green. Benches carry
pre-existing graph-Value API drift (unrelated, out of scope).
