--- src/api_handlers/request_handlers_tests.rs
@@ -1,663 +1,0 @@
-//! Basic tests for UnifiedHandlers to establish initial test coverage
-//!
-//! These tests focus on the core functionality and public APIs of the unified
-//! handlers, including collection operations, vector operations, and error handling.
-
-#[cfg(test)]
-mod tests {
-    use super::super::request_handlers::*;
-    use crate::proto::proximadb_v1::{
-        CollectionConfig, CollectionOperation, CollectionRequest, IncludeFields, SearchQuery,
-        VectorBatchRequest, VectorOperation, VectorSearchRequest,
-    };
-    use chrono::Utc;
-    use std::collections::HashMap;
-
-    /// Helper to create test collection config
-    fn create_test_collection_config(name: &str) -> CollectionConfig {
-        use crate::proto::proximadb_v1::{DistanceMetric, StorageEngine};
-
-        CollectionConfig {
-            name: name.to_string(),
-            dimension: 128,
-            distance_metric: Some(DistanceMetric::Cosine as i32),
-            storage_engine: Some(StorageEngine::Sst as i32),
-            filterable_columns: vec![],
-            index_configs: vec![],
-            quantization: None,
-            description: Some(format!("Test collection {}", name)),
-            tags: vec!["test".to_string()],
-            storage_config: None,
-            embedding_models: vec![],
-            primary_index: Some("hnsw".to_string()),
-            auto_index_selection: Some(true),
-            owner: Some("test_user".to_string()),
-            record_schema: None,
-            enable_proxima_record: None,
-            text_columns: vec![],
-            text_storage_configs: vec![],
-            enable_dual_use_embeddings: None,
-            canonical_embedding_precision: None,
-            permitted_principals: vec![],
-            index_policy: None,
-            pax_vector_quant: None,
-        }
-    }
-
-    /// Helper to create canonical test vector record.
-    fn create_test_vector_record(id: &str, dimension: usize) -> proximadb_records::ProximaRecord {
-        create_test_proxima_record(
-            id,
-            (0..dimension).map(|i| i as f32 * 0.1).collect(),
-            Utc::now().timestamp(),
-        )
-        .with_record_version(1)
-    }
-
-    fn create_test_proxima_record(
-        id: &str,
-        vector: Vec<f32>,
-        timestamp_ms: i64,
-    ) -> proximadb_records::ProximaRecord {
-        let timestamp_ns = timestamp_ms.saturating_mul(1_000_000);
-        let dim = vector.len() as u32;
-        proximadb_records::ProximaRecord {
-            oid: id.to_string(),
-            created_at_ns: timestamp_ns,
-            updated_at_ns: timestamp_ns,
-            embeddings: vec![proximadb_records::EmbeddingCell {
-                model_id: "default".to_string(),
-                modality: "dense_vector".to_string(),
-                dim,
-                values: proximadb_records::EmbeddingValues::Fp32(vector),
-                ..Default::default()
-            }],
-            ..proximadb_records::ProximaRecord::default()
-        }
-    }
-
-    trait TestProximaRecordExt {
-        fn with_record_version(self, version: u64) -> proximadb_records::ProximaRecord;
-        fn with_valid_to_ms(self, expires_at_ms: i64) -> proximadb_records::ProximaRecord;
-    }
-
-    impl TestProximaRecordExt for proximadb_records::ProximaRecord {
-        fn with_record_version(mut self, version: u64) -> proximadb_records::ProximaRecord {
-            self.record_version = version;
-            self
-        }
-
-        fn with_valid_to_ms(mut self, expires_at_ms: i64) -> proximadb_records::ProximaRecord {
-            self.valid_to_ns = Some(expires_at_ms.saturating_mul(1_000_000));
-            self
-        }
-    }
-
-    #[test]
-    fn test_sql_query_result_structure() {
-        // Test SqlQueryResult struct creation and field access
-        let rows = vec![
-            serde_json::json!({"id": "1", "name": "test1"}),
-            serde_json::json!({"id": "2", "name": "test2"}),
-        ];
-        let columns = vec![
-            ("id".to_string(), "string".to_string()),
-            ("name".to_string(), "string".to_string()),
-        ];
-
-        let result = SqlQueryResult {
-            rows: rows.clone(),
-            columns: columns.clone(),
-            row_count: 2,
-            execution_time_ms: 100,
-        };
-
-        assert_eq!(result.rows.len(), 2);
-        assert_eq!(result.columns.len(), 2);
-        assert_eq!(result.row_count, 2);
-        assert_eq!(result.columns[0].0, "id");
-        assert_eq!(result.columns[1].0, "name");
-
-        // Test Debug implementation
-        let debug_str = format!("{:?}", result);
-        assert!(debug_str.contains("SqlQueryResult"));
-    }
-
-    #[test]
-    fn test_collection_config_creation() {
-        use crate::proto::proximadb_v1::{DistanceMetric, StorageEngine};
-
-        // Test collection config creation helper
-        let config = create_test_collection_config("test_collection");
-
-        assert_eq!(config.name, "test_collection");
-        assert_eq!(config.dimension, 128);
-        assert_eq!(config.distance_metric, Some(DistanceMetric::Cosine as i32));
-        assert_eq!(config.storage_engine, Some(StorageEngine::Sst as i32));
-        assert!(config.description.is_some());
-        assert!(!config.tags.is_empty());
-        assert!(config.filterable_columns.is_empty());
-        assert!(config.index_configs.is_empty());
-    }
-
-    #[test]
-    fn test_vector_record_creation() {
-        // Test vector record creation helper
-        let record = create_test_vector_record("test_id", 128);
-
-        assert_eq!(record.oid, "test_id");
-        assert_eq!(record.embeddings[0].values.len(), 128);
-        assert_eq!(record.record_version, 1);
-        assert!(record.created_at_ns > 0);
-        assert!(record.props.is_empty());
-    }
-
-    #[test]
-    fn test_collection_operations_enum_values() {
-        // Test that CollectionOperation enum has expected values
-        let create_op = CollectionOperation::CollectionCreate as i32;
-        let get_op = CollectionOperation::CollectionGet as i32;
-        let list_op = CollectionOperation::CollectionList as i32;
-        let update_op = CollectionOperation::CollectionUpdate as i32;
-        let delete_op = CollectionOperation::CollectionDelete as i32;
-
-        // These should be different values
-        assert_ne!(create_op, get_op);
-        assert_ne!(get_op, list_op);
-        assert_ne!(list_op, update_op);
-        assert_ne!(update_op, delete_op);
-    }
-
-    #[test]
-    fn test_vector_operations_enum_values() {
-        // Test that VectorOperation enum has expected values
-        let batch_op = VectorOperation::VectorBatch as i32;
-        let search_op = VectorOperation::VectorSearch as i32;
-        let get_op = VectorOperation::VectorGet as i32;
-
-        // These should be different values
-        assert_ne!(batch_op, search_op);
-        assert_ne!(search_op, get_op);
-    }
-
-    #[test]
-    fn test_search_query_creation() {
-        // Test SearchQuery structure
-        let query = SearchQuery {
-            vector: vec![0.1, 0.2, 0.3, 0.4],
-            filters: HashMap::new(),
-            advanced_filter: None,
-        };
-
-        assert_eq!(query.vector.len(), 4);
-        assert!(query.filters.is_empty());
-        assert!(query.advanced_filter.is_none());
-    }
-
-    #[test]
-    fn test_include_fields_structure() {
-        // Test IncludeFields structure
-        let fields = IncludeFields {
-            vector: true,
-            metadata: false,
-            score: true,
-            rank: true,
-            source: false,
-            source_options: HashMap::new(),
-        };
-
-        assert!(fields.vector);
-        assert!(!fields.metadata);
-        assert!(fields.score);
-        assert!(fields.rank);
-    }
-
-    #[test]
-    fn test_collection_request_structure() {
-        // Test CollectionRequest structure
-        let config = create_test_collection_config("test");
-        let request = CollectionRequest {
-            operation: CollectionOperation::CollectionCreate as i32,
-            collection_id: Some("test_collection".to_string()),
-            collection_config: Some(config),
-            query_params: HashMap::new(),
-            options: HashMap::new(),
-            migration_config: HashMap::new(),
-        };
-
-        assert_eq!(
-            request.operation,
-            CollectionOperation::CollectionCreate as i32
-        );
-        assert_eq!(request.collection_id.unwrap(), "test_collection");
-        assert!(request.collection_config.is_some());
-    }
-
-    #[test]
-    fn test_vector_batch_request_structure() {
-        // Test VectorBatchRequest structure
-        let vectors: Vec<crate::proto::proximadb_v1::VectorRecord> = vec![
-            create_test_vector_record("vec1", 128),
-            create_test_vector_record("vec2", 128),
-        ]
-        .into_iter()
-        .map(Into::into)
-        .collect();
-
-        let request = VectorBatchRequest {
-            collection_id: "test_collection".to_string(),
-            vectors,
-        };
-
-        assert_eq!(request.collection_id, "test_collection");
-        assert_eq!(request.vectors.len(), 2);
-    }
-
-    #[test]
-    fn test_vector_search_request_structure() {
-        // Test VectorSearchRequest structure
-        let query = SearchQuery {
-            vector: vec![0.1, 0.2, 0.3, 0.4],
-            filters: HashMap::new(),
-            advanced_filter: None,
-        };
-
-        let request = VectorSearchRequest {
-            collection_id: "test_collection".to_string(),
-            queries: vec![query],
-            top_k: 10,
-            distance_metric_override: None,
-            search_params: None,
-            include_fields: Some(IncludeFields {
-                vector: true,
-                metadata: true,
-                score: true,
-                rank: true,
-                source: false,
-                source_options: HashMap::new(),
-            }),
-            search_optimization: None,
-        };
-
-        assert_eq!(request.collection_id, "test_collection");
-        assert_eq!(request.queries.len(), 1);
-        assert_eq!(request.top_k, 10);
-        assert!(request.include_fields.is_some());
-    }
-
-    #[test]
-    fn test_vector_record_field_access() {
-        // Test the v1 boundary projection from canonical ProximaRecord.
-        let record: crate::proto::proximadb_v1::VectorRecord =
-            create_test_proxima_record("test_id", vec![1.0, 2.0, 3.0], 1234567890)
-                .with_record_version(2)
-                .with_valid_to_ms(1234567999)
-                .into();
-
-        assert_eq!(record.id, "test_id");
-        assert_eq!(record.vector.len(), 3);
-        assert_eq!(record.timestamp, Some(1234567890));
-        // Note: created_at is now timestamp in proto VectorRecord
-        assert_eq!(record.updated_at, Some(1234567890));
-        assert_eq!(record.expires_at.unwrap(), 1234567999);
-        assert_eq!(record.version, Some(2));
-    }
-
-    #[test]
-    fn test_collection_config_optional_fields() {
-        // Test CollectionConfig with different optional field configurations
-        use crate::proto::proximadb_v1::{DistanceMetric, StorageEngine};
-
-        let mut config = CollectionConfig {
-            name: "test".to_string(),
-            dimension: 256,
-            distance_metric: Some(DistanceMetric::Euclidean as i32),
-            storage_engine: Some(StorageEngine::Sst as i32),
-            filterable_columns: vec![],
-            index_configs: vec![],
-            quantization: None,
-            description: None,
-            tags: vec![],
-            storage_config: None,
-            embedding_models: vec![],
-            primary_index: Some("hnsw".to_string()),
-            auto_index_selection: Some(true),
-            owner: Some("test_user".to_string()),
-            record_schema: None,
-            enable_proxima_record: None,
-            text_columns: vec![],
-            text_storage_configs: vec![],
-            enable_dual_use_embeddings: None,
-            canonical_embedding_precision: None,
-            permitted_principals: vec![],
-            index_policy: None,
-            pax_vector_quant: None,
-        };
-
-        assert_eq!(config.dimension, 256);
-        assert_eq!(
-            config.distance_metric,
-            Some(DistanceMetric::Euclidean as i32)
-        );
-        assert!(config.description.is_none());
-        assert!(config.tags.is_empty());
-
-        // Test setting optional fields
-        config.description = Some("Updated description".to_string());
-
-        assert_eq!(config.description.unwrap(), "Updated description");
-    }
-
-    #[test]
-    fn test_multiple_search_queries() {
-        // Test VectorSearchRequest with multiple queries
-        let queries = vec![
-            SearchQuery {
-                vector: vec![0.1, 0.2],
-                filters: HashMap::new(),
-                advanced_filter: None,
-            },
-            SearchQuery {
-                vector: vec![0.3, 0.4],
-                filters: HashMap::new(),
-                advanced_filter: None,
-            },
-            SearchQuery {
-                vector: vec![0.5, 0.6],
-                filters: HashMap::new(),
-                advanced_filter: None,
-            },
-        ];
-
-        let request = VectorSearchRequest {
-            collection_id: "multi_query_collection".to_string(),
-            queries,
-            top_k: 5,
-            distance_metric_override: None,
-            search_params: None,
-            include_fields: None,
-            search_optimization: None,
-        };
-
-        assert_eq!(request.queries.len(), 3);
-        assert_eq!(request.top_k, 5);
-    }
-
-    #[test]
-    fn test_vector_batch_with_different_dimensions() {
-        // Test creating vectors with different dimensions
-        let vectors: Vec<crate::proto::proximadb_v1::VectorRecord> = vec![
-            create_test_vector_record("vec_64", 64),
-            create_test_vector_record("vec_128", 128),
-            create_test_vector_record("vec_256", 256),
-        ]
-        .into_iter()
-        .map(Into::into)
-        .collect();
-
-        assert_eq!(vectors[0].vector.len(), 64);
-        assert_eq!(vectors[1].vector.len(), 128);
-        assert_eq!(vectors[2].vector.len(), 256);
-
-        // Each vector should have expected value pattern
-        assert_eq!(vectors[0].vector[0], 0.0);
-        assert_eq!(vectors[0].vector[1], 0.1);
-        assert_eq!(vectors[1].vector[10], 1.0);
-        assert_eq!(vectors[2].vector[20], 2.0);
-    }
-
-    #[test]
-    fn test_collection_tags_handling() {
-        // Test collection config with tags
-        use crate::proto::proximadb_v1::{DistanceMetric, StorageEngine};
-
-        let config = CollectionConfig {
-            name: "metadata_test".to_string(),
-            dimension: 100,
-            distance_metric: Some(DistanceMetric::Cosine as i32),
-            storage_engine: Some(StorageEngine::Sst as i32),
-            filterable_columns: vec![],
-            index_configs: vec![],
-            quantization: None,
-            description: Some("Collection with tags".to_string()),
-            tags: vec!["production".to_string(), "v1".to_string()],
-            storage_config: None,
-            embedding_models: vec![],
-            primary_index: Some("hnsw".to_string()),
-            auto_index_selection: Some(true),
-            owner: Some("test_user".to_string()),
-            record_schema: None,
-            enable_proxima_record: None,
-            text_columns: vec![],
-            text_storage_configs: vec![],
-            enable_dual_use_embeddings: None,
-            canonical_embedding_precision: None,
-            permitted_principals: vec![],
-            index_policy: None,
-            pax_vector_quant: None,
-        };
-
-        assert_eq!(config.tags.len(), 2);
-        assert!(config.tags.contains(&"production".to_string()));
-        assert!(config.tags.contains(&"v1".to_string()));
-    }
-
-    #[test]
-    fn test_request_operation_type_conversion() {
-        // Test converting between different operation types
-        let operations = vec![
-            (CollectionOperation::CollectionCreate, "create"),
-            (CollectionOperation::CollectionGet, "get"),
-            (CollectionOperation::CollectionList, "list"),
-            (CollectionOperation::CollectionUpdate, "update"),
-            (CollectionOperation::CollectionDelete, "delete"),
-        ];
-
-        for (op, _name) in operations {
-            let op_value = op as i32;
-            assert!(op_value >= 0);
-            // Test that each operation has a unique integer value
-            let unique_ops: std::collections::HashSet<i32> = vec![
-                CollectionOperation::CollectionCreate as i32,
-                CollectionOperation::CollectionGet as i32,
-                CollectionOperation::CollectionList as i32,
-                CollectionOperation::CollectionUpdate as i32,
-                CollectionOperation::CollectionDelete as i32,
-            ]
-            .into_iter()
-            .collect();
-            assert_eq!(unique_ops.len(), 5); // All should be unique
-        }
-    }
-
-    #[test]
-    fn test_edge_case_vector_values() {
-        // Test vector with edge case values
-        let edge_vector = vec![
-            0.0,       // Zero
-            1.0,       // One
-            -1.0,      // Negative
-            0.000001,  // Very small positive
-            -0.000001, // Very small negative
-            999999.0,  // Very large positive
-            -999999.0, // Very large negative
-        ];
-
-        let record: crate::proto::proximadb_v1::VectorRecord = create_test_proxima_record(
-            "edge_case_vector",
-            edge_vector.clone(),
-            Utc::now().timestamp(),
-        )
-        .with_record_version(1)
-        .into();
-
-        assert_eq!(record.vector.len(), 7);
-        assert_eq!(record.vector[0], 0.0);
-        assert_eq!(record.vector[1], 1.0);
-        assert_eq!(record.vector[2], -1.0);
-        assert!(record.vector[3].abs() < 0.00001);
-        assert!(record.vector[4] < 0.0);
-        assert!(record.vector[5] > 999000.0);
-        assert!(record.vector[6] < -999000.0);
-    }
-
-    #[test]
-    fn test_request_timeout_values() {
-        // Test different timeout configurations
-        let timeouts = vec![
-            None,         // No timeout
-            Some(1000),   // 1 second
-            Some(30000),  // 30 seconds
-            Some(300000), // 5 minutes
-        ];
-
-        for timeout in timeouts {
-            let vectors: Vec<crate::proto::proximadb_v1::VectorRecord> =
-                vec![create_test_vector_record("test", 64)]
-                    .into_iter()
-                    .map(Into::into)
-                    .collect();
-            let _request = VectorBatchRequest {
-                collection_id: "timeout_test".to_string(),
-                vectors,
-            };
-
-            // Remove timeout assertions as batch_timeout_ms field doesn't exist
-            let _timeout = timeout; // Prevent unused variable warning
-        }
-    }
-
-    #[test]
-    fn test_vector_batch_configurations() {
-        // Test different vector batch configurations
-        let batch_sizes = vec![
-            None,     // Default batch size
-            Some(1),  // Single vector
-            Some(3),  // Small batch
-            Some(5),  // Medium batch
-            Some(10), // Large batch
-        ];
-
-        for batch_size in batch_sizes {
-            // VectorBatchRequest doesn't have batch_size field, so we test with vectors count instead
-            let vectors_count = batch_size;
-            let mut vectors = Vec::new();
-            if let Some(count) = vectors_count {
-                for i in 0..count.min(10) {
-                    // Limit to 10 for test performance
-                    let record: crate::proto::proximadb_v1::VectorRecord =
-                        create_test_vector_record(&format!("test_{}", i), 32).into();
-                    vectors.push(record);
-                }
-            }
-
-            let request = VectorBatchRequest {
-                collection_id: "batch_size_test".to_string(),
-                vectors: vectors.clone(),
-            };
-
-            assert_eq!(request.vectors.len(), vectors.len());
-        }
-    }
-
-    /// Test hybrid search API integration
-    #[tokio::test]
-    async fn test_hybrid_search_api() {
-        use crate::core::search::hybrid::FusionStrategy;
-
-        // This test demonstrates the hybrid search API
-        // In production, this would be called from REST/gRPC endpoints
-
-        let collection_id = "test_collection";
-        let text_query = "machine learning algorithms";
-        let query_vector = [0.1, 0.2, 0.3, 0.4, 0.5];
-        let top_k = 10;
-        let _fusion_strategy = FusionStrategy::ReciprocalRank { k: 60 };
-
-        // Note: This would require a full UnifiedSearchHandler instance
-        // For now, we're testing that the API compiles and the types work together
-
-        // Verify fusion strategy can be created
-        let _strategy = FusionStrategy::ReciprocalRank { k: 60 };
-
-        // Verify parameters are valid
-        assert_eq!(collection_id, "test_collection");
-        assert!(!text_query.is_empty());
-        assert_eq!(query_vector.len(), 5);
-        assert_eq!(top_k, 10);
-
-        println!("✅ Hybrid search API compiles and integrates correctly");
-    }
-
-    // ============================================================
-    // CollectionIdCache unit tests (coverage improvement)
-    // ============================================================
-
-    #[test]
-    fn test_collection_id_cache_new() {
-        let cache = CollectionIdCache::new();
-        assert!(cache.get("nonexistent").is_none());
-    }
-
-    #[test]
-    fn test_collection_id_cache_insert_and_get() {
-        let cache = CollectionIdCache::new();
-        cache.insert("my_collection".to_string(), "uuid-123".to_string());
-        assert_eq!(cache.get("my_collection"), Some("uuid-123".to_string()));
-    }
-
-    #[test]
-    fn test_collection_id_cache_miss() {
-        let cache = CollectionIdCache::new();
-        cache.insert("exists".to_string(), "uuid-1".to_string());
-        assert!(cache.get("does_not_exist").is_none());
-    }
-
-    #[test]
-    fn test_collection_id_cache_invalidate() {
-        let cache = CollectionIdCache::new();
-        cache.insert("col_a".to_string(), "id-a".to_string());
-        cache.insert("col_b".to_string(), "id-b".to_string());
-
-        cache.invalidate("col_a");
-        assert!(cache.get("col_a").is_none());
-        assert_eq!(cache.get("col_b"), Some("id-b".to_string()));
-    }
-
-    #[test]
-    fn test_collection_id_cache_invalidate_by_value() {
-        let cache = CollectionIdCache::new();
-        cache.insert("my_name".to_string(), "target-id".to_string());
-        cache.invalidate("target-id");
-        assert!(cache.get("my_name").is_none());
-    }
-
-    #[test]
-    fn test_collection_id_cache_ttl_expiry() {
-        use std::time::Duration;
-        let cache = CollectionIdCache::with_ttl(Duration::from_millis(50));
-        cache.insert("ephemeral".to_string(), "uuid-x".to_string());
-        assert!(cache.get("ephemeral").is_some());
-
-        std::thread::sleep(Duration::from_millis(60));
-        assert!(cache.get("ephemeral").is_none());
-    }
-
-    #[test]
-    fn test_collection_id_cache_overwrite() {
-        let cache = CollectionIdCache::new();
-        cache.insert("col".to_string(), "old-id".to_string());
-        cache.insert("col".to_string(), "new-id".to_string());
-        assert_eq!(cache.get("col"), Some("new-id".to_string()));
-    }
-
-    #[test]
-    fn test_collection_id_cache_multiple_entries() {
-        let cache = CollectionIdCache::new();
-        for i in 0..50 {
-            cache.insert(format!("col_{}", i), format!("id_{}", i));
-        }
-        assert_eq!(cache.get("col_0"), Some("id_0".to_string()));
-        assert_eq!(cache.get("col_49"), Some("id_49".to_string()));
-    }
-} // mod tests
