+++ src/network/rest/v1/handlers.rs
@@ -77,8 +77,8 @@ impl RestCoreServices {
             document_service: services.document_service.clone(),
             collection_service: services.collection_service.clone(),
             graph_operations_service: services.graph_service.clone(),
-            graph_collection_service: services.request_handlers.graph_collection_service.clone(),
-            record_ops: services.request_handlers.record_ops(),
+            graph_collection_service: services.graph_collection_service.clone(),
+            record_ops: services.record_ops.clone(),
             observability_service: services.observability_service.clone(),
             event_log: services.event_log.clone(),
             api_handlers: services.api_handlers.clone(),
